Restore new pcm-telechips.c with DMA
In 4417a79, firmware/target/arm/pcm-telechips.c switched to a much
older version, 9ee9cf7, for unknown reasons. It lacks DMA and other
optimizations. The new version with DMA works, but hasn't been
tested as extensively. It is being introduced with the assumption
that the pcm-telechips.c change in 4417a79 was a mistake.

Change-Id: I7d589f0deef570cf50ea6798b47e42b5fcc7c82a